Convert NatService generator to use dep injection
Rather than getting NatService generator to call out to
the EdgeGateway for each interface, utilise the
Core::EdgeGateway.interfaces method to supply an array
of EdgeGatewayInterface objects to it, which can then be
used to look up the needed network details.

In the process, adjusted the interface of NatService to take
the input config as the primary argument, and the interfaces
list as the secondary.

The additional code has ensured that the Interface: sections
in the output vCloud have a mime 'type:' value. This was always present
in the API, but not handled in the unit tests - and hence the addition
of these fields in the tests.