Don't reset edge gateway after integration tests
Currently, we reset the edge gateway configuration both before and after
the integration tests are run.

Because the edge gateway is reset before the tests are run, it should
not be necessary to reset it afterwards. Furthermore, this extends the
time it takes for the integration tests to complete as evidenced by the
'VCLOUD_NO_RESET_VSE_AFTER' environment variable which this commit
removes.